Treat all errors encountered while patching as fatal
This will stop the reconcile. No further patches, resources, or
functions will be processed.

This commit breaks required field paths, in that we'll return a fatal
result if a required field path is missing. This means a set of patches
that could have been eventually consistent now cannot. I intend to
address that in a following commit.

Signed-off-by: Nic Cope <nicc@rk0n.org>