Don't delete resources when required field paths don't exist
If we hit a composition error that is likely due to a misconfigured
composition we just return a fatal error. Previously we'd skip the
affected composed resource, which could result in it being deleted if it
already existed.

We'll now only skip rendering a composed resource if a required field
path is not present _and_ the composed resource doesn't exist yet.

Signed-off-by: Nic Cope <nicc@rk0n.org>