Remove RenderJSON function
This function was copied from c/c but isn't needed here.

RenderJSON tried to preserve metadata of the passed object (e.g. name
and namespace), and checked that its GVK didn't change. In practice in
this function the passed object will always be new and empty so we can
just use json.Marshal directly.

Signed-off-by: Nic Cope <nicc@rk0n.org>